+++ com.nuttty.eureka.order/src/main/java/com/nuttty/eureka/order/application/fegin/CompanyClient.java
@@ -1,6 +1,6 @@
-package com.nuttty.eureka.order.infrastructure.fegin;
+package com.nuttty.eureka.order.application.fegin;
 
-import com.nuttty.eureka.order.infrastructure.fegin.dto.CompanyInfoDto;
+import com.nuttty.eureka.order.application.fegin.dto.CompanyInfoDto;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
